chore(Benchmark): rename e2e → statistic_build_probe
