+++ scripts/benchmarking/statistic_build_probe/query-configs/AccuracyGtPoint_ClusterMonitoring.yaml.template
