+++ scripts/benchmarking/statistic_build_probe/query-configs/AccuracyGtPoint_Nexmark.yaml.template
