+++ scripts/benchmarking/statistic_build_probe/query-configs/AccuracyGtRange_ClusterMonitoring.yaml.template
