+++ scripts/benchmarking/statistic_build_probe/query-configs/CountMinAccuracyApproxPoint_ClusterMonitoring.yaml.template
