+++ scripts/benchmarking/statistic_build_probe/query-configs/CountMinAccuracyApproxPoint_Nexmark.yaml.template
