+++ scripts/benchmarking/statistic_build_probe/query-configs/CountMinBuild_Nexmark.yaml.template
