+++ scripts/benchmarking/statistic_build_probe/query-configs/CountMinProbe_Nexmark.yaml.template
