+++ scripts/benchmarking/statistic_build_probe/query-configs/EquiWidthHistogramAccuracyApprox_ClusterMonitoring.yaml.template
