+++ scripts/benchmarking/statistic_build_probe/query-configs/EquiWidthHistogramAccuracyApprox_Nexmark.yaml.template
