+++ scripts/benchmarking/statistic_build_probe/query-configs/EquiWidthHistogramBuild_ClusterMonitoring.yaml.template
