+++ scripts/benchmarking/statistic_build_probe/query-configs/EquiWidthHistogramBuild_Nexmark.yaml.template
