+++ scripts/benchmarking/statistic_build_probe/query-configs/EquiWidthHistogramProbe_ClusterMonitoring.yaml.template
