+++ scripts/benchmarking/statistic_build_probe/query-configs/PassthroughBuild_ClusterMonitoring.yaml.template
