+++ scripts/benchmarking/statistic_build_probe/query-configs/ReservoirAccuracyApprox_ClusterMonitoring.yaml.template
