+++ scripts/benchmarking/statistic_build_probe/query-configs/ReservoirBuild_ClusterMonitoring.yaml.template
