+++ scripts/benchmarking/statistic_build_probe/query-configs/ReservoirBuild_Nexmark.yaml.template
