+++ scripts/benchmarking/statistic_build_probe/query-configs/ReservoirProbe_ClusterMonitoring.yaml.template
