+++ scripts/benchmarking/statistic_build_probe/query-configs/SumBuild_ClusterMonitoring.yaml.template
