+++ scripts/benchmarking/statistic_build_probe/query-configs/SumBuild_Nexmark.yaml.template
